+++ resonos/SQL/resonos/tag.sql
@@ -20,10 +20,10 @@ SELECT * FROM tag WHERE id IS NULL;
 
 -- 더미 데이터 
 INSERT INTO `tag` (`id`, `name`) VALUES
-  (1, '밝음'),
-  (2, '어둠'),
-  (3, '신남'),
-  (4, '잔잔함'),
+  -- (1, '밝음'),
+  -- (2, '어둠'),
+  -- (3, '신남'),
+  -- (4, '잔잔함'),
   (5, '몽환'),
   (6, '감성'),
   (7, '파워풀');
