+++ resonos/src/main/java/com/cosmus/resonos/service/BadgeService.java
@@ -24,7 +24,7 @@ public interface BadgeService {
     // 미회득 배지 조회
     public List<Badge> doesNotHaveBadge(Long loginUserId) throws Exception;
     // 배지 소유 여부 체크
-    public boolean checkBadge(Long loginUserId, Long badgeId) throws Exception;
+    public boolean checkBadge(@Param("loginUserId") Long loginUserId, @Param("badgeId") Long badgeId) throws Exception;
     // 최근 획득 배지
     public List<Badge> recentGetBadge(Long id) throws Exception;
     // 획득 배지 수 조회
