+++ resonos/src/main/resources/com/cosmus/resonos/mapper/PolicyMapper.xml
@@ -20,7 +20,7 @@
     SELECT * FROM policy WHERE id = #{id}
   </select>
 
-  <insert id="insert" useGeneratedKeys="true">
+  <insert id="insert" >
     INSERT INTO policy (type, content, version, updated_at)
     VALUES (#{type}, #{content}, #{version}, NOW())
   </insert>
